Request to start 2013-12-30-0 to 2014-01-16-0
The modelio.nml files in this commit list are redundant
in CaseDocs and BuildConf/cplconf.  In addition, they
were created by the last cycle of the previous job,
and will be overwritten by the first cycle of the next job.
So they have almost no useful information to review.
Should I add one or both sets to the .gitignore file
to clarify what people should pay attention to?
Or I could add just the *0080 andor *0001 files to the commit
if anyone is worried about documenting everything that changes.

Someone should look at
/glade/scratch/raeder/f.e21.FHIST_BGC.f09_025.CAM6assim.011/run,
where I've reset the case to start from 2013-12-30-00000.
That did not require visible changes to the files in $CASEROOT,
so git will not bring them to your attention.

If you're feeling obsessive you can check the math
about the number of cycles specified, compared to the
intended run dates.  The cycles are the definitive number,
but it's good to be consistent between plans and intentions.

In these early approvals, it would be helpful to know what
you looked at.