Set up archive of run environments: CASEROOT/logs.
This first version is from the last cycle of Rean_spinup_2010,
which was before the June 2019 "upgrade".  It will help show
the environment changes due to the upgrade.  Subsequent versions
of this file will include the LIDs of all the cycles of a whole
month, in addition to the environment.  Only selected months
will be archived: months from years that didn't need to be rerun,
and before and after system down times.