adding dry-run ability
logs run-time options into the output file.
untested, since it has to be run in a CASEROOT, and
I don't have permission to run in Kevin's CASEROOT.
Only purges one month, and it has error-checking to ensure it
cannot be the month that is currently running (i.e. partially complete).